Boom OpenGL!
It was fun and frustrating, getting GPU support to this thing. I insanely searched a bug that crashes when mouse moves. It appeared to be a code that trying to use CPU specific struct to get mouse's position while on GPU specific render. Boom. Some wrong place reaching in ram problem. Meant to be crash.